pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.8",
    "numpy>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
